16.2.prototype Versus __proto__{proto and prototype?}
proto and prototype?

Ok, so how does this relate to the prototype key in our code?. Whenever you use the new keyword, anything on the prototype key of that object becomes the __proto__ of the newly instantiated object.

Ok, let's play with this!

Create a new variable called playObject set to an object with a lift key set to stuff. Next, say playObject.__proto__.cat = 'meow'. You shouldn't normally access or set the __proto__ property directly, but for playing around now, it's great. Finally, console.log(playObject.lift), which we know will work, but also playObject.cat

Ok, try it. Refresh! Hey, stuff and meow! That's the __proto__ property in action!