18.2.AJAX Form Submit: The Lazy Way{Making $wrapper Wrap Everything}
Making $wrapper Wrap Everything

In both situations, step one is the same: we need attach a listener on submit of the form. Head over to our template.
The form itself lives in another template that's included here: _form.html.twig inside app/Resources/views/lift.
This is a Symfony form, but all this fanciness ultimately renders a good, old-fashioned form tag. Give the form another class: js-new-rep-log-form.
Copy that and head into RepLogApp so we can attach a new listener. But wait... there is one problem: the $wrapper is actually the <table> element.
And the form does not live inside of the <table>!

When you create little JavaScript applications like RepLogApp, you want the wrapper to be an element that goes around everything you need to manipulate.

Ok, no problem: let's move the js-rep-log-table class from the table itself to the div that surrounds everything.
Down below, I don't need to change anything here, but let's rename $table to $wrapper for clarity.