18.3.AJAX Form Submit: The Lazy Way{The Form Submit Listener}
The Form Submit Listener

Now adding our listener is simple: this.$wrapper.find() and look for .js-new-rep-log-form. Then, .on('submit'), have this call a new method: this.handleNewFormSubmit. And don't forget the all-important .bind(this)

Down below, add that function - handleNewFormSubmit - and give it the event argument. This time, calling e.preventDefault() will prevent the form from actually submitting, which is good. For now, just console.log('submitting')

Ok, test time! Head back, refresh, and try the form. Yes! We get the log, but the form doesn't submit.