18.4.AJAX Form Submit: The Lazy Way{Adding AJAX}
Adding AJAX

Turning this form into an AJAX call will be really easy... because we already know that this form works if we submit it in the traditional way. So let's just literally send that exact same request, but via AJAX.

First, get the form with $form = $(e.currentTarget). Next, add $.ajax(), set the url to $form.attr('action') and the method to POST. For the data, use $form.serialize().

That's a really lazy way to get all the values for all the fields in the form and put them in the exact format that the server is accustomed to seeing for a form submit.

That's already enough to work! Submit that form! Yea, you can see the AJAX calls in the console and web debug toolbar. Of course, we don't see any new rows until we manually refresh the page...

So that's where the real work starts: showing the validation errors on the form on error and dynamically inserting a new row on success. Let's do it