19.4.Old-School AJAX HTML{Distinguishing Between Success and Error}
Distinguishing Between Success and Error

Our JavaScript code needs to be able to distinguish between a successful request and one that failed with validation errors.

There's a perfectly standard way of doing this... and I was being lazy until now! On error, we should not return a 200 status code, and that's what the render() function gives us by default. When you return a 200 status code, it activates jQuery's success handler.

Instead, we should return a 400 status code, or really, anything that starts with a 4. To do that, add $html = and then change render() to renderView()

As soon as we do that, the success function will not be called on errors. Instead, the error function will be called. For an error callback, the first argument is not the data from the response, it's a jqXHR object. That's fine, because the response content is stored on jqXHR.responseText.

Now we can use the success function to add the new tr to the table. Before the AJAX call - to avoid any problems with the this variable - add $tbody = this.$wrapper.find('tbody'). And in success, add $tbody.append(data)

That should do it!

Try it! Refresh the page! If we submit with errors, we get the errors! If we submit with something correct, a new row is added to the table. The only problem is that it doesn't update the total dynamically - that still requires a refresh.