20.1.Delegate Selectors FTW!
So dang. Each time we submit, it adds a new row to the table, but its delete button doesn't work until we refresh. What's going on here?

Well, let's think about it. In RepLogApp, the constructor function is called when we instantiate it. So, inside $(document).ready():

 67 lines  app/Resources/views/lift/index.html.twig
{% block javascripts %}
    <script>
        $(document).ready(function() {
            var $wrapper = $('.js-rep-log-table');
            var repLogApp = new RepLogApp($wrapper);
        });
    </script>
{% endblock %}
That means it's executed after the entire page has loaded.

Then, at that exact moment, our code finds all elements with a js-delete-rep-log class in the HTML, and attaches the listener to each DOM Element:

 99 lines  web/assets/js/RepLogApp.js
(function(window, $) {
    window.RepLogApp = function ($wrapper) {
        this.$wrapper.find('.js-delete-rep-log').on(
            'click',
            this.handleRepLogDelete.bind(this)
        );
    };
})(window, jQuery);
So if we have 10 delete links on the page initially, it attaches this listener to those 10 individual DOM Elements. If we add a new js-delete-rep-log element later, there will be no listener attached to it. So when we click delete, nothing happens! So, what's the fix?

If you're like me, you've probably fixed this in a really crappy way before. Back then, after dynamically adding something to my page, I would manually try to attach whatever listeners it needed. This is SUPER error prone and annoying