21.3.Proper JSON API Endpoint Setup{Updating the AJAX Call}
Updating the AJAX Call

Ok! Let's update our AJAX call to go to this endpoint. In RepLogApp, down in handleNewFormSubmit, we need to somehow get that URL
No problem! Find the form and add a fancy new data-url attribute set to path(), then the name of that route: rep_log_new
Bam! Now, back in RepLogApp, before we use that, let's clear out all the code that actually updates our DOM: all the stuff related to updating the form with the form errors or adding the new row. That's all a todo for later
But, do add a console.log('success') and console.log('error') so we can see if this stuff is working!
Finally, update the url to $form.data('url'). Next, our data format needs to change - I'll show you exactly how.