23.2.Handling JSON Validation Errors{Creating a Selectors Map}
Creating a Selectors Map

In _mapErrorsToForm, let's look at the error data and use it to add an error span below that field. Obviously, we need to use jQuery to find our .js-new-rep-log-form form element.

But wait! Way up in our constructor, we're already referencing this selector.
It's no big deal, but I would like to not duplicate that class name in multiple places. Instead, add an _selectors property to our object. Give it a newRepForm key that's set to its selector. Now, reference that with this._selectors.newRepForm. Now, reference that with this._selectors.newRepForm
Below in our function, do the same: var $form = this.$wrapper.find(this._selectors.newRepForm)