23.3.Handling JSON Validation Errors{Mapping the Data Manually}
Mapping the Data Manually

Now what? Simple: loop over every field see if that field's name is present in the errorData. And if it is, add an error message span element below the field. To find all the fields, use $form.find(':input') - that's jQuery magic to find all form elements. Then, .each() and pass it a callback function. Inside, we know that this is actually the form element. So we can say var fieldName = $(this).attr('name')
I'm also going to find the wrapper that's around the entire form field. What I mean is, each field is surrounded by a .form-group element. Since we're using Bootstrap, we also need to add a class to this. Find it with var $wrapper = $(this).closest('.form-group')
Perfect!

Then, if there is not any data[fieldName], the field doesn't have an error. Just continue.
If there is an error, we need to add some HTML to the page. The easy way to do that is by creating a new jQuery element. Set var $error to $() and then the HTML you want: a span with a js-field-error class and a help-block class. I left the span blank because it's cleaner to add the text on the next line: $error.html(errorsData[fieldName])
This jQuery object is now done! But it's not on the page yet.