23.4.Handling JSON Validation Errors{Mapping the Data Manually}
Add it with $wrapper.append($error). Also call $wrapper.addClass('has-error').
Yes! Let's try it! Refresh and submit! There it is!

The only problem is that, once I finally fill in the field, the error message stays! AND, I get a second error message! Man, we gotta get this thing cleaned up