23.5.Handling JSON Validation Errors{Mapping the Data Manually}
We gotta get this thing cleaned up!
No problem: at the top, use $form.find() to find all the .js-field-error elements. And, remove those. Next, find all the form-group elements and remove the has-error class
Refresh now, and re-submit! Errors! Fill in one... beautiful!
And if we fill in both fields, the AJAX call is successful, but nothing updates. Time to tackle that.