24.1.Clearing the Form, Prepping for a Template
Let's do the easy thing first: when we submit the form successfully, these errors need to disappear!
We already have code for that, so copy it, and isolate it into its own new method called _removeFormErrors. Call that from our map function.
The other thing we should do is empty, or reset the fields after submit. Let's create another function that does that and removes the form's errors. Call it _clearForm. First call this._removeFormErrors(). To "reset" the form, get the DOM Element itself - there will be only one - by using [0] and calling reset() on it. I love that this [0] thing isn't a mystery anymore! Call this from up in success: self._clearForm()
Ok, test this baby out! Submit it empty, then fill it out for real and submit. Boom