26.2.Full-JavaScript Rendering & FOSJsRoutingBundle{composer require friendsofsymfony/jsrouting-bundle}
Getting the /reps URL

But first, the question is: how can we get this URL inside of JavaScript? I mean, we could hardcode it, but that should be your last option. Well, I can think of three ways:

1.We could add a data- attribute to something, like on the $wrapper element in index.html.twig.

2.We could pass the URL into our RepLogApp object via a second argument to the constructor, just like we're doing with .

3.If you're in Symfony, you could cheat and use a cool library called FOSJsRoutingBundle.

Using FOSJsRoutingBundle

Google for that, and click the link on the Symfony.com documentation. This allows you to expose some of your URLs in JavaScript. Copy the composer require line, open up a new tab, paste that and hit enter:

composer require friendsofsymfony/jsrouting-bundle