26.4.Full-JavaScript Rendering & FOSJsRoutingBundle{Using FOSJsRoutingBundle}
Now, head to RepLogController. In order to make this route available to that Routing JavaScript variable, we need to add options={"expose" = true}
Back in RepLogApp, remember that this library gives us a global Routing object. And of course, inside of our self-executing function, we do have access to global variables. But as a best practice, we prefer to pass ourselves any global variables that we end up using. So at the bottom, pass in the global Routing object, and then add Routing as an argument on top