28.3.Catching a Failed Promise {Catch Recovers from Errors}
Catch Recovers from Errors

catch is named catch for a reason: you really need to think about it in the same way as a try-catch block in PHP. It will catch the failed Promise above and return a new Promise that resolves successfully. That means that any handlers attached to it - like our second .then() - will execute as if everything was fine.
We're going to talk more about this, but obviously, this is probably not what we want. Instead, move the .catch() to the end

Now, the second .then() will only be executed if the first .then() is executed. The .catch() will catch any failed Promises - or errors - at the bottom. More on the error catching later.

Refresh now! Cool - only the catch() handler is running.