29.1.Promise catch: Catches Errors? {Making our Endpoint Less Friendly; A catch Catches Errors}
Yay! Let's complicate things!

Our AJAX call works really well, because when we make an AJAX call to create a new RepLog, our server returns all the data for that new RepLog. That means that when we call .then() on the AJAX promise, we have all the data we need to call _addRow() and get that new row inserted!

Too easy: so let's make it harder!

Making our Endpoint Less Friendly

Pretend that we don't have full control over our API. And instead of returning the RepLog data from the create endpoint - which is what this line does - it returns an empty response
Passing null means no response content, and 204 is just a different status code used for empty responses - that part doesn't make any difference. Now head over and fill out the form successfully. Whoa!
Yep, it blew up - that's not too surprising: we get an error that says:
  totalWeightLifted is not defined.

And if you look closely, that's coming from underscore.js. This is almost definitely an error in our template. We pass the response data - which is now empty - into ._addRow() .And that eventually becomes the variables for the template
An empty response means that no variables are being passed. Hence, totalWeightLifted is not defined.
But check this out: there's a second error:
  JSON Exception: unexpected token

A catch Catches Errors

This is coming from RepLogApp.js, line 89. Woh, it's coming from inside our .catch() handler. Here's the deal: in reality, .catch() will be called if your Promise is rejected, or if a handler above it throws an error. Since our .then() calls _addRow() and that throws an exception, this ultimately triggers the .catch(). Again, this works a lot like the try-catch block in PHP!

 Tip

There are some subtle cases when throwing an exception inside asynchronous code won't trigger your .catch(). The Mozilla Promise Docs(https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/catch#Gotchas_when_throwing_errors) discuss this!