29.2.Promise catch: Catches Errors? {Conditionally Handling in catch}
Conditionally Handling in catch

So this complicates things a bit. Before, we assumed that the value passed to .catch() would always be the jqXHR object: that's what jQuery passes when its Promise is rejected. But now, we're realizing that it might not be that, because something else might fail.

Let's console.log(jqXHR). Ok, refresh and fill out our form. There it is! Thanks to the error, it logs a "ReferenceError".

We've just found out that .catch() will catch anything that went wrong... and that the value passed to your handler will depend on what went wrong.