29.4.Promise catch: Catches Errors? {Conditionally Handling in catch}
What JavaScript doesn't have is the ability to do more intelligent try-catch block, where you catch only certain types of errors. Instead, .catch() handles all errors, but then, you can write your code to be a bit smarter.

Since we really only want to catch jqXHR errors, we could check to see if the jqXHR value is what we're expecting. One way is to check if jqXHR.responseText === 'undefined'. If this is undefined, this is not the error we intended to handle. To not handle it, and make that error uncaught, just throw jqXHR. Now, if you wanted to, you could add another .catch() on the bottom, and inside its function, log the e value.
You see, because the first catch throws the error, the second one will catch it.
And when we try it now, the error prints two times - jQuery's Promise logs a warning each time an error is thrown inside a Promise. And then at the bottom, there's our log.