29.5.Promise catch: Catches Errors? {Conditionally Handling in catch}
Let's remove the second .catch() and the if statement
Why? Well, I'm not going to code defensively unless I'm coding against a situation that might possibly happen. In this case, it was developer error: my code just isn't written correctly for the server. Instead of trying to code around that, we just need to fix things!
We do the same thing in PHP: most of the time, we let exceptions happen... because it means we messed up!
Ok, we understand more about .catch(), but we still need to fix this whole situation! To do that, we'll need to create our own Promise.