30.1.Making (and Keeping) a Promise
Ignore the error for a second and go down to the AJAX call. We know that this method returns a Promise, and then we call .then() on it:
171 lines  web/assets/js/RepLogApp.js
(function(window, $, Routing) {
    $.extend(window.RepLogApp.prototype, {
        handleNewFormSubmit: function(e) {
            this._saveRepLog(formData)
            .then(function(data) {
            });
        },
        _saveRepLog: function(data) {
            return $.ajax({
                url: Routing.generate('rep_log_new'),
                method: 'POST',
                data: JSON.stringify(data)
            });
        },
    });
})(window, jQuery, Routing);
But, our handler expects that the Promise's value will be the RepLog data. But now, it's null because that's what the server is returning!

 132 lines  src/AppBundle/Controller/RepLogController.php
class RepLogController extends BaseController
{
    public function newRepLogAction(Request $request)
    {
        //$response = $this->createApiResponse($apiModel);
        $response = new Response(null, 204);
    }
}

Somehow, I want to fix this method so that it once again returns a Promise whose value is the RepLog data.

How? Well first, we're going to read the Location header that's sent back in the response - which is the URL we can use to fetch that RepLog's data:

 132 lines  src/AppBundle/Controller/RepLogController.php
class RepLogController extends BaseController
{
    public function newRepLogAction(Request $request)
    {
        // setting the Location header... it's a best-practice
        $response->headers->set(
            'Location',
            $this->generateUrl('rep_log_get', ['id' => $repLog->getId()])
        );
    }
}
We'll use that to make a second AJAX call to get the data we need.