30.2.Making (and Keeping) a Promise {Making the Second AJAX Call}
Start simple: add another .then() to this, with 3 arguments: data, textStatus and jqXHR.
Normally, promise handlers are only passed 1 argument, but in this case jQuery cheats and passes us 3. To fetch the Location header, say console.log(jqXHR.getResponseHeader('Location'))
Go see if that works: we still get the errors, but hey! It prints /reps/76!