30.3.Making (and Keeping) a Promise {Making the Second AJAX Call; Could we use a Promise?}
Let's make an AJAX call to that: copy the jqXHR line. Then, add our favorite $.ajax() and set the URL to that header. Add a .then() to this Promise with a data argument.
Finally, this should be the RepLog data.
To check things, add console.log('now we are REALLY done') and also console.log(data) to make sure it looks right.
Ok, refresh and fill out the form. Ignore the errors, because there's our message and the correct data!
Ok, now we can just return this somehow, right? Wait, that's not going to work... When we return the main $.ajax(), that Promise is resolved - meaning finished - the moment that the first AJAX call is made. You can see that because the errors from the handlers happen first, and then the second AJAX call finishes.
Somehow, we need to return a Promise that isn't resolved until that second AJAX call finishes.
There are two ways to do this - we'll do the harder way... because it's a lot more interesting - but I'll mention the other way at the end.

Could we use a Promise?

What we need to do is create our own Promise object, and take control of exactly when it's resolved and what value is passed back.
If you look at the Promise documentation, you'll find an example of how to do this: new Promise() with one argument: a function that has resolve and reject arguments. I know, it looks a little weird.
Inside of that function, you'll put your asynchronous code. And as soon as it's done, you'll call the resolve() function and pass it whatever value should be passed to the handlers. If something goes wrong, call the reject() function. This is effectively what jQuery is doing right now inside of its $.ajax() function.