30.4.Making (and Keeping) a Promise {Browser Compatability! Polyfill}
Browser Compatability! Polyfill

There's one quick gotcha: not all browsers support the Promise object. But, no worries! Google for "JavaScript Promise polyfill CDN".
A polyfill is a library that gives you functionality that's normally only available in a newer version of your language, JavaScript in this case. PHP also has polyfills: small PHP libraries that backport newer PHP functionality.
This polyfill guarantees that the Promise object will exist in JavaScript. If it's already supported by the browser it uses that. But if not, it adds it.
Copy the es6-promise.auto.min.js path. In the next tutorial, we'll talk all about what that es6 part means. Next, go into app/Resources/views/base.html.twig and add a script tag with src="" and this path.
Now our Promise object is guaranteed!